+++ tests/TestPackageData/app/Events/OrderPlaced.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace NyonCode\LaravelPackageToolkit\Tests\TestPackageData\app\Events;
+
+class OrderPlaced
+{
+    public function __construct(public string $id = '') {}
+}